+++ onediff_diffusers_extensions/onediffx/__init__.py
@@ -1,7 +1,5 @@
 __version__ = "0.13.0.dev"
-from onediff.infer_compiler.oneflow_compiler_config import (
-    oneflow_compiler_config as compiler_config,
-)
+from onediff.infer_compiler.oneflow import oneflow_compiler_config as compiler_config
 from .compilers.diffusion_pipeline_compiler import compile_pipe, save_pipe, load_pipe
 
 __all__ = ["compile_pipe", "compiler_config", "save_pipe", "load_pipe"]
