+++ src/onediff/infer_compiler/deployable_module.py
@@ -1,0 +1,10 @@
+from typing import Any
+import torch
+
+
+class DeployableModule(torch.nn.Module):
+    def __init__(self):
+        torch.nn.Module.__init__(self)
+
+    def forward(self, *args, **kwargs) -> Any:
+        raise NotImplementedError()
