+++ src/onediff/infer_compiler/oneflow/__init__.py
@@ -1,0 +1,1 @@
+from .config import OneFlowCompilerConfig, oneflow_compiler_config
